Move translation config to development environment
When sites are translated their pages will all be served at URLs that
are prefixed with the locale. There's no need for most of the tests to
be concerned with translations, so it doesn't make sense to run them all
in the context of a translated site.

We therefore benefit from making the translation config for spec/dummy
environment specific.