Navigation links use localized URLs
On translated sites, paths in URLs are prefixed with the locale (e.g.
/de/permalink). Links in the primary and secondary nav to other pages on
the site need to preserve the locale by linking to /de/foo and /de/bar.

This change maintains the locale by linking to a page's
`localized_path`, rather than its `path`.

The patch also removes the `minimal` scope that was used to instantiate
some "dehydrated" instances of the pages that are required to generate
the HTML for the primary and secondary nav.

While `minimal` was a nice performance improvement, it doesn't work with
localization fallbacks (it supports localization, just not the
fallbacks).

To provide a bit of context on fallbacks, when Mongoid retrieves the
value of a localized field it first checks the value for the current
locale. If the field doesn't have a value defined in that locale, it can
check one of several "fallbacks" for that locale. Taking the 'de' locale
as an example, 'en' is defined as a fallback.

We're depending on this fallback behaviour in Slices; when the value of
a field is requested in a locale for which no translation exists,
Mongoid will usually return the value for the 'en' locale (as 'en' is
defined as a fallback for many other locales).

The `minimal` scope was implemented using the `.only` method, which is
defined in the Origin gem (Origin is Mongoid's query builder).
Unfortunately the `Origin::Smash#normalized_key` method converts the
`name` key into `name.de` (by appending the current locale) before the
query is sent to Mongo. We therefore only retrieve the value for the
current locale, and are unable to "fall back" onto the value for any
other locales.

If we continued to use the `minimal` scope we'd find that translated
sites could have links to untranslated pages in their navigation whose
"link text" was the path (i.e. <a ...>/de/path</a>). Given that the
navigation HTML is cached, it seems better to generate accurate HTML a
bit more slowly, than to generate HTML with seemingly broken links in.

Removing `minimal` was discussed with Erin.