Stub out `Slices::InstallGenerator#optionally_create_mongoid_yaml`
It calls `generate` which shells out to `script/rails`. The way the generator/specs are set up means it doesn't exist: the generator expects to be run in `Rails.root` (so in this case, because Slices is an engine, in `spec/dummy`). But if you run it there the spec ends up deleting everything (because it calls `prepare_destination`; hence why `destination` is set to `Rails.root.join('tmp')`). I couldn't find an answer within the time box I gave myself.