Add a scope for loading pages without their slices
Sometimes we just want the page metadata itself, so no point taking the performance hit of loading slices too. Sometimes we use `only` for this, but the problem with that is that it's not obvious that a field has not been loaded when you access it - you just get nil. `excluding_slices` is a compromise - it should be pretty clear when slices haven't been loaded, and the small risk of it not is outweighed by the performance benefits.